[function-sig-opts] Extract FunctionSignatureTransform into FunctionSignatureTransform.h in preparation for exploding out function-sig-opts.
I am also going to use this to implement Guaranteed -> Owned in its own file.
The exploding of the other parts of function-sig-opts will happen later.

rdar://38196046
